Struct histogram::Config

pub struct Config {
    // some fields omitted
}

A configuration struct for building custom Histograms.

Methods

impl Config

fn new() -> Config

create a new Histogram Config with defaults

Example

let mut c = Histogram::configure();

fn precision(self, precision: u32) -> Self

set HistogramConfig precision

Example

let mut c = Histogram::configure();
c.precision(4); // set to 4 significant figures

fn max_memory(self, bytes: u32) -> Self

set HistogramConfig memory limit

Example

let mut c = Histogram::configure();
c.max_memory(1024*1024); // cap Histogram at 1MB of data

fn max_value(self, max: u64) -> Self

set HistogramConfig value limit

Example

let mut c = Histogram::configure();
c.max_value(1000); // values above 1000 will not be stored

fn build(self) -> Option<Histogram>

Build a new histogram based on the current configuration values. Return None if the new histogram would require more than the configured memory size.